Moved generic functions in gobblin-aws module to GobblinAWSUtils
